выполнил задание модуляция
